redo state management functions, remove some logic from reducer functions
