+++ src/components/NavComponents/MainNav.tsx
@@ -1,7 +1,7 @@
 import { useBoardDataContext } from '../../contexts/StateManagement'
 import { EditType } from '../../interfaces/DataInterfaces'
 import { MainNavProps } from '../../interfaces/PropInterfaces'
-import { extractColumns } from '../../utilities/dataManipulation'
+import { extractColumns } from '../../utilities/dataUtilities'
 import ButtonPrimary from '../Buttons/ButtonPrimary'
 import Edit from '../Edit'
 import NavDark from '../icons/NavDark'
