+++ src/contexts/StateManagement.tsx
@@ -10,7 +10,7 @@ import {
   TaskInterface,
 } from '../interfaces/DataInterfaces'
 import { populateBoardObject } from '../utilities/populateBoard'
-
+import { deleteBoard, deleteColumn } from '../utilities/dataUtilities'
 const BoardDataContext = createContext<BoardData | null>(null)
 const BoardDispatchContext = createContext<React.Dispatch<BoardAction> | null>(
   null
@@ -75,7 +75,7 @@ function reducer(
       const { id, title } = action.payload as BoardInterface
       return {
         ...state,
-        boards: { ...state.boards, [action.payload.id]: { id, title } },
+        boards: { ...state.boards, [id]: { id, title } },
       }
     }
 
@@ -88,8 +88,7 @@ function reducer(
     case DATA_ACTION.DELETE_BOARD: {
       const { id } = action.payload as BoardInterface
       let stateCopy = structuredClone(state)
-      stateCopy = { ...deleteAssociatedColumns(id, stateCopy) }
-      delete stateCopy.boards[id]
+      stateCopy = deleteBoard(id, stateCopy)
 
       return { ...stateCopy, activeBoard: Object.keys(stateCopy.boards)[0] }
     }
@@ -105,10 +104,9 @@ function reducer(
     case DATA_ACTION.DELETE_COLUMN: {
       const { id } = action.payload as ColumnInterface
       let stateCopy = structuredClone(state)
-      stateCopy = deleteAssociatedTasks(id, state)
-      delete stateCopy.columns[id]
+      stateCopy = deleteColumn(id, stateCopy)
 
-      return { ...stateCopy }
+      return stateCopy
     }
 
     case DATA_ACTION.UPDATE_TASK: {
@@ -135,50 +133,4 @@ function reducer(
     default:
       return state
   }
-}
-
-export function deleteAssociatedSubtasks(
-  taskId: string,
-  data: BoardData
-): BoardData {
-  const clonedData = structuredClone(data)
-  const { subtasks } = clonedData
-
-  for (let [key, subtask] of Object.entries<SubtaskInterface>(subtasks)) {
-    if (subtask.taskId === taskId) delete subtasks[key]
-  }
-
-  return { ...clonedData }
-}
-
-export function deleteAssociatedTasks(
-  columnId: string,
-  data: BoardData
-): BoardData {
-  let clonedData = structuredClone(data)
-  const { tasks } = clonedData
-  for (let [key, task] of Object.entries<TaskInterface>(tasks)) {
-    if (task.columnId === columnId) {
-      clonedData = { ...deleteAssociatedSubtasks(key, clonedData) }
-      delete clonedData.tasks[key]
-    }
-  }
-
-  return { ...clonedData }
-}
-
-export function deleteAssociatedColumns(
-  boardId: string,
-  data: BoardData
-): BoardData {
-  let clonedData = structuredClone(data)
-  const { columns } = clonedData
-  for (let [key, column] of Object.entries<ColumnInterface>(columns)) {
-    if (column.boardId === boardId) {
-      clonedData = { ...deleteAssociatedTasks(key, clonedData) }
-      delete clonedData.columns[key]
-    }
-  }
-
-  return { ...clonedData }
 }
